+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+* Authorizing a key for a single bucket ensures that this bucket is cached
 * `Bucket.ls` operation supports wildcard matching strings
 
 ### Infrastructure
+* Additional tests for listing files/versions
 * Ensured that changelog validation only happens on pull requests
 * Upgraded GitHub actions checkout to v3, python-setup to v4